pyproject.toml
[project]
name = "blacksmith"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"

[project.scripts]
bots = 'bot_models.async_main:run_system'

[tool.poetry]
packages = [{ include = "bot_models", from = "src" },
           { include = "operations", from = "src" },
           { include = "configs", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13.0"
pandas="^2.2.0"
matplotlib="^3.10.0"
binance-connector="^3.12.0"
python-binance="^1.0.0"
rich="^14.0.0"

[tool.poetry.group.dev.dependencies]
requests="^2.32.0"
black="^25.1.0"
mypy="^1.15.0"
pylint="^3.3.0"
loguru="^0.7.0"
numba = "^0.61.2"
nest-asyncio = "^1.6.0"

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
